+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add missing documentation for `single_statement_only` style of `RSpec/ImplicitSubject` cop. ([@tejasbubane][])
 * Fix an exception in `DescribedClass` when accessing a constant on a variable in a spec that is nested in a namespace. ([@rrosenblum][])
 * Add new `RSpec/IdenticalEqualityAssertion` cop. ([@tejasbubane][])
+* Add `RSpec/Rails/AvoidSetupHook cop. ([@paydaylight][])
 
 ## 2.3.0 (2021-04-28)
 
@@ -618,3 +619,4 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@stephannv]: https://github.com/stephannv
 [@Tietew]: https://github.com/Tietew
 [@rrosenblum]: https://github.com/rrosenblum
+[@paydaylight]: https://github.com/paydaylight