+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Master (Unreleased)
 
+* Update `RSpec/FilePath` to check suffix when given a non-constant top-level node (e.g. features). ([@topalovic][])
 * Add missing documentation for `single_statement_only` style of `RSpec/ImplicitSubject` cop. ([@tejasbubane][])
 * Fix an exception in `DescribedClass` when accessing a constant on a variable in a spec that is nested in a namespace. ([@rrosenblum][])
 * Add new `RSpec/IdenticalEqualityAssertion` cop. ([@tejasbubane][])
@@ -621,3 +622,4 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@Tietew]: https://github.com/Tietew
 [@rrosenblum]: https://github.com/rrosenblum
 [@paydaylight]: https://github.com/paydaylight
+[@topalovic]: https://github.com/topalovic